+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Change 266 files from CRLF to LF.
 - Run tests on push as well as on a pull request so developers can see impact
 - `ArduinoBackend` now exposes `config_file_path` instead of `config_dir` so that we can be explicit about [strange behavior in `arduino-cli` that isn't going to change anytime soon](https://github.com/arduino/arduino-cli/issues/753)
+- Use `arduino-cli` version `0.29.0` as the backend
 
 ### Deprecated